requirements.txt
kaldi_native_fbank
mediapipe
tqdm
scikit-learn
glfw
PyOpenGL